services/rasp_car_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pymysql
cryptography
paho-mqtt
dotenv